+++ src/test/java/domain/betting/BettingAmountTest.java
@@ -2,13 +2,14 @@ package domain.betting;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import java.math.BigDecimal;
 import org.junit.jupiter.api.Test;
 
 class BettingAmountTest {
 
     @Test
     void 베팅_금액이_음수면_예외가_발생한다() {
-        Integer testBettingAmount = -1000;
+        BigDecimal testBettingAmount = BigDecimal.valueOf(-1000);
         assertThrows(IllegalArgumentException.class, () -> {
             new BettingAmount(testBettingAmount);
         });
@@ -16,7 +17,7 @@ class BettingAmountTest {
 
     @Test
     void 베팅_금액이_양수이면_생성된다() {
-        Integer testBettingAmount = 1000;
+        BigDecimal testBettingAmount = BigDecimal.valueOf(1000);
         assertDoesNotThrow(() -> {
             new BettingAmount(testBettingAmount);
         });
